requirements.txt
streamlit==1.31.1
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
requests==2.31.0
google-api-python-client==2.101.0